+++ CHANGELOG.md
@@ -6,6 +6,13 @@ CHANGELOG
 
 ### Latest changes
 
+* VFS-5599 Added support for Onepanel Proxy URL
+* VFS-5605 Updated list of space privileges
+* VFS-5607 Updated lodash in onedata-gui-websocket-client
+* VFS-5402 Updated lodash in onedata-gui-common
+* VFS-5502 Fixed not working generate new token button on Oneprovider deploy views
+* VFS-5402 Updated common libraries
+* VFS-5519 Fixed not visible block with service version in Safari
 * VFS-5476 Refactoring of token/origin endpoints and unfied GUI URL
 * VFS-5219 Data discovery functionality
 * VFS-5425 Service Pack 3 for unified GUI